move pages to different folders
